pyproject.toml
[tool.poetry]
name = "template-pytorch-lightning-hydra-mlflow-poetry"
version = "0.1.0"
description = ""
authors = ["hppRC <hpp.ricecake@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
# please lookup a compatible PyTorch version with your OS and CUDA from:  https://download.pytorch.org/whl/torch_stable.html
torch = {url = "https://download.pytorch.org/whl/cu111/torch-1.9.0%2Bcu111-cp37-cp37m-linux_x86_64.whl"}
tqdm = "^4.61.2"
pytorch-lightning = "^1.3.8"
mlflow = "^1.19.0"
hydra-colorlog = "^1.1.0"
hydra-core = "^1.1.0"
omegaconf = "^2.1.0"

[tool.poetry.dev-dependencies]
pysen = {version = "^0.9.1", extras = ["lint"]}
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pysen]
version = "0.9"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = false
mypy_preset = "strict"
py_version = "py37"
[[tool.pysen.lint.mypy_targets]]
paths = ["."]
